Use the same property makefile variable as the phones we're overriding
Problem:
A change updated the makefile variable we're using to set system
properties from PRODUCT_PRODUCT_PROPERTIES to
PRODUCT_PROPERTY_OVERRIDES, which had the impact of our car phone builds
failing to replace the ?= defined defaults that the phone supplied such
that we currently enable nearly all Bluetooth profiles.

Solution:
We can use the same makefile variable as the phones we're overriding to
allow their defaults to be replaced with our desired values.

Bug: 318386540
Test: m
Test: adb shell getprop | grep -ie bluetooth
Change-Id: Ida5e4173b42fb8772cd9cb738f168e955baa47eb